vm_voting on migration action are sometimes not getting rejected
Due to race conditions, multiple messages might be received from a
single read by guestServer. guestServer in this case would only
handle the first message and discard the remaining ones.

In this particular issue, guestServer received a heartbeat challenge
response message and a vote notification response (reject) message
from a single read, and the latter message was discarded.

This fix re-writes message handler for virtio serial channel to
handle segmented and multiple messages, and removed any newline
characters in client log message.

It contains fix for both host and client, so both host load and
guest image need to be updated to test the fix.

Change-Id: I50ed2236ad24b61a79dce7773ace5b4b6ee33f28
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>